Try setting Milvus2 as the default for tests
