+++ setup.cfg
@@ -112,11 +112,11 @@ only-faiss-gpu =
 faiss-gpu = 
     farm-haystack[sql,only-faiss-gpu]
 only-milvus1 = 
-    pymilvus<2.0.0  # Refer milvus version support matrix at https://github.com/milvus-io/pymilvus#install-pymilvus
+    pymilvus<2.0.0
 milvus1 = 
     farm-haystack[sql,only-milvus1]
 only-milvus = 
-    pymilvus>=2.0.0  # Refer milvus version support matrix at https://github.com/milvus-io/pymilvus#install-pymilvus
+    pymilvus>=2.0.0
 milvus = 
     farm-haystack[sql,only-milvus]
 weaviate =
@@ -124,9 +124,9 @@ weaviate =
 graphdb = 
     SPARQLWrapper
 docstores =
-    farm-haystack[faiss,milvus1,weaviate,graphdb]
+    farm-haystack[faiss,milvus,weaviate,graphdb]
 docstores-gpu =
-    farm-haystack[faiss-gpu,milvus1,weaviate,graphdb]
+    farm-haystack[faiss-gpu,milvus,weaviate,graphdb]
 crawler = 
     selenium
     webdriver-manager
